Add public proxy option helpers
