+++ src/Handler/CurlFactory.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Exception\TimeoutException;
 use GuzzleHttp\Promise as P;
 use GuzzleHttp\Promise\PromiseInterface;
+use GuzzleHttp\ProxyOptions;
 use GuzzleHttp\Psr7\FnStream;
 use GuzzleHttp\Psr7\LazyOpenStream;
 use GuzzleHttp\TransferStats;
@@ -899,61 +900,6 @@ class CurlFactory implements CurlFactoryInterface
     }
 
     /**
-     * @return array{set: bool, proxy: string, no_proxy: string, effective: ?string}
-     */
-    private static function resolveProxy(RequestInterface $request, array $options): array
-    {
-        $resolved = [
-            'set' => false,
-            'proxy' => '',
-            'no_proxy' => '',
-            'effective' => null,
-        ];
-
-        if (!isset($options['proxy'])) {
-            return $resolved;
-        }
-
-        $proxy = $options['proxy'];
-        if (!\is_array($proxy)) {
-            if (!\is_string($proxy)) {
-                throw new \InvalidArgumentException('proxy must be a string or array');
-            }
-
-            $resolved['set'] = true;
-            $resolved['proxy'] = $proxy;
-            $resolved['effective'] = $proxy !== '' ? $proxy : null;
-
-            return $resolved;
-        }
-
-        $scheme = $request->getUri()->getScheme();
-        if (!isset($proxy[$scheme])) {
-            return $resolved;
-        }
-
-        if (!\is_string($proxy[$scheme])) {
-            throw new \InvalidArgumentException('proxy values must be strings');
-        }
-
-        $uri = $request->getUri();
-        $noProxy = isset($proxy['no']) ? Utils::normalizeNoProxy($proxy['no']) : [];
-        if ($noProxy !== [] && Utils::isUriInNoProxy($uri, $noProxy)) {
-            $resolved['set'] = true;
-            $resolved['proxy'] = '';
-            $resolved['no_proxy'] = '*';
-
-            return $resolved;
-        }
-
-        $resolved['set'] = true;
-        $resolved['proxy'] = $proxy[$scheme];
-        $resolved['effective'] = $proxy[$scheme] !== '' ? $proxy[$scheme] : null;
-
-        return $resolved;
-    }
-
-    /**
      * @return array<int|string, mixed>
      */
     private function getDefaultConf(EasyHandle $easy): array
@@ -986,8 +932,8 @@ class CurlFactory implements CurlFactoryInterface
                 throw new \RuntimeException('HTTP/3 is not supported by this cURL installation.');
             }
 
-            $proxy = self::resolveProxy($easy->request, $easy->options);
-            $conf[\CURLOPT_HTTP_VERSION] = null !== $proxy['effective']
+            $proxy = ProxyOptions::resolve($easy->request->getUri(), $easy->options['proxy'] ?? null);
+            $conf[\CURLOPT_HTTP_VERSION] = $proxy->hasProxy()
                 ? (CurlVersion::supportsHttp2() ? \CURL_HTTP_VERSION_2_0 : \CURL_HTTP_VERSION_1_1)
                 : (int) \constant('CURL_HTTP_VERSION_3');
         } elseif ('2' === $version || '2.0' === $version) {
@@ -1257,11 +1203,14 @@ class CurlFactory implements CurlFactoryInterface
             $conf[\CURLOPT_NOSIGNAL] = true;
         }
 
-        $proxy = self::resolveProxy($easy->request, $options);
-        $selectedProxy = $proxy['effective'];
-        if ($proxy['set']) {
-            $conf[\CURLOPT_PROXY] = $proxy['proxy'];
-            $conf[\CURLOPT_NOPROXY] = $proxy['no_proxy'];
+        $proxy = ProxyOptions::resolve($easy->request->getUri(), $options['proxy'] ?? null);
+        $selectedProxy = $proxy->getProxy();
+        if ($selectedProxy !== null) {
+            $conf[\CURLOPT_PROXY] = $selectedProxy;
+            $conf[\CURLOPT_NOPROXY] = '';
+        } elseif ($proxy->shouldDisableProxy()) {
+            $conf[\CURLOPT_PROXY] = '';
+            $conf[\CURLOPT_NOPROXY] = $proxy->isBypassed() ? '*' : '';
         }
 
         $proxyForConnectionReuse = self::getEffectiveProxyForConnectionReuse($selectedProxy, $options);
